dagger 2 refactoring

+++ domain/src/main/java/com/free/domain/Coord.kt
@@ -1,6 +1,6 @@
 package com.free.domain
 
 data class Coord(
-    val lat: Int,
-    val lon: Int
+    val lat: Double,
+    val lon: Double
 )
